requirements.txt
pillow==10.2.0
pip-autoremove==0.10.0
pipdeptree==2.14.0
